// src/app/interfaces/users.ts
export class Users {

}

export interface IUser {
  id?: string;
  username?: string;
  email?: string;
  plan?: string;
  verified?: boolean;
  password?: string;
}

export interface IDatos {
  psid?: string;
  value?: string;
  pid?: string;
  type?: string;
  email?: string;
  like?: string;
  dislike?: string;
}